extract constructibles
